+++ server/db/models/customer.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   const Customer = sequelize.define('Customer', {
     customername: DataTypes.STRING,
-    nohp: DataTypes.INTEGER,
+    nohp: DataTypes.BIGINT,
     email: DataTypes.STRING,
     tanggal_lahir: DataTypes.DATE,
     noktp: DataTypes.INTEGER,
